requirements.txt
# Deep Learning
# NOTE: Install PyTorch separately with CUDA support using:
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
# For CPU only: pip3 install torch torchvision torchaudio
# Do NOT install via requirements.txt to ensure correct CUDA version
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0

# Vector Database
qdrant-client>=1.7.0

# Financial Data
polygon-api-client>=1.12.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
scikit-learn>=1.3.0

# Technical Analysis
ta>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
